feat: Integrate watchlist API and update UI (#46)

+++ frontend/utils/database/favorites/removeFromFavorites.ts
@@ -1,8 +1,8 @@
 "use server";
 import { createClient } from "@/utils/supabase/server";
-import { FAVORITES_TABLE } from "../constants";
+import { FAVORITES_TABLE, MediaType } from "../../constants";
 
-async function removeFromFavorites(mediaType: string, mediaId: string) {
+async function removeFromFavorites(mediaType: MediaType, mediaId: string) {
   const supabase = createClient();
   const {
     data: { user },
@@ -15,6 +15,7 @@ async function removeFromFavorites(mediaType: string, mediaId: string) {
   const { data, error } = await supabase
     .from(FAVORITES_TABLE)
     .delete()
+    .eq("user_id", user.id)
     .eq("media_type", mediaType)
     .eq("media_id", mediaId);
